move GraphNodeType to nodes
